requirements.txt
requests
pytest
pytest-xdist
playwright
python-dotenv
allure-pytest